requirements.txt
python-telegram-bot[ext]
psycopg2-binary # For PostgreSQL later
